pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "costs"
version = "0.1.41"
description = "Zero-config AI cost calculator per commit/model with liteLLM"
readme = "README.md"
license = "Apache-2.0"
authors = [
    {name = "Tom Sapletta", email = "tom@sapletta.com"}
]
requires-python = ">=3.9"
dependencies = [
    "gitpython>=3.1",
    "pandas>=2.0",
    "typer>=0.12",
    "click<8.1.0",
    "httpx>=0.27",
    "litellm>=1.0",
    "python-dotenv>=1.0"
]

[project.scripts]
costs = "costs.cli:main"

[tool.poetry]
packages = [{include = "costs", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
gitpython = "^3.1"
pandas = "^2.0"
typer = "^0.12"
httpx = "^0.27"
litellm = "^1.0"
python-dotenv = "^1.0"

[tool.poetry.scripts]
costs = "costs.cli:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
black = "^24.0"

[tool.costs]
# Badge generation configuration
badge = true  # Enable automatic badge generation
badge_format = "svg"  # Badge format: svg, png
badge_style = "flat"  # Badge style: flat, flat-square, plastic, for-the-badge
badge_label = "AI Cost"  # Badge label text
update_readme = true  # Auto-update README.md with badge
readme_path = "README.md"  # Path to README file

# Cost thresholds for badge colors (in USD)
badge_color_thresholds = { low = 1.0, medium = 5.0, high = 10.0, critical = 50.0 }

# Analysis configuration
default_model = "openrouter/qwen/qwen3-coder-next"
analysis_mode = "byok"  # byok, local, saas
full_history = false  # Analyze full git history by default
max_commits = 1000  # Maximum commits to analyze

# Auto-run configuration
# Set to true to enable automatic badge generation on package install
auto_generate = false

# Badge output directory (relative to project root)
badge_output_dir = "."

# Custom badge message template
# Available variables: {cost}, {model}, {commits}, {date}
badge_message_template = "${cost} with {model} · {commits} AI commits"

[tool.pfix]
# Self-healing Python configuration
model = "openrouter/qwen/qwen3-coder-next"
auto_apply = true
auto_install_deps = true
auto_restart = false
max_retries = 3
create_backups = false
git_auto_commit = false

[tool.pfix.runtime_todo]
enabled = true
todo_file = "TODO.md"
min_severity = "low"
deduplicate = true
